Resolve TODO by moving some functions into a common base class #346
